+++ src/hint_processor/builtin_hint_processor/builtin_hint_processor_definition.rs
@@ -80,7 +80,7 @@ use felt::Felt252;
 #[cfg(feature = "skip_next_instruction_hint")]
 use crate::hint_processor::builtin_hint_processor::skip_next_instruction::skip_next_instruction;
 
-use super::vrf::pack_512::inv_mod_p_uint512;
+use super::vrf::inv_mod_p_uint512::inv_mod_p_uint512;
 
 pub struct HintProcessorData {
     pub code: String,
