+++ src/hint_processor/builtin_hint_processor/vrf/mod.rs
@@ -1,1 +1,1 @@
-pub mod pack_512;
+pub mod inv_mod_p_uint512;
